Edits to both Server & Client
Shows Like/Dislike counter, there is an error that is stopping the client-side from initializing. I'm not sure what's causing it, because it appears to me to have gone from working to broken with no change to the code (obviously there must have been changes, I just can't find them).
I deleted the beds, because I was having issues wrapping my head around them and finally just decided it'd be best -for me- to just delete them and work on getting the flowers to show up and spend time on getting the like/dislike counter to work properly, which it still is not.
... I realized there should have been more incremental commits... Sorry!